+++ src/main/java/com/study/petory/common/exception/GlobalExceptionHandler.java
@@ -1,8 +1,9 @@
-package com.study.petory.exception;
+package com.study.petory.common.exception;
 
 import java.util.HashMap;
 import java.util.Map;
 
+import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.HttpMessageNotReadableException;
@@ -10,11 +11,8 @@ import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
+import com.study.petory.common.exception.enums.ErrorCode;
 import com.study.petory.common.response.CommonResponse;
-import com.study.petory.exception.enums.ErrorCode;
 
 import lombok.extern.slf4j.Slf4j;
